@remotion/studio: Split selected outline UV controls
